Created new component for plugins list inside app modal changed the copyright for plugins section inside app modal
