+++ frontend/src/_components/AppModal.jsx
@@ -5,8 +5,7 @@ import { ButtonSolid } from '@/_ui/AppButton/AppButton';
 import _, { noop } from 'lodash';
 import { validateName } from '@/_helpers/utils';
 import { FormWrapper } from './FormWrapper';
-import SolidIcon from '@/_ui/Icon/SolidIcons';
-import config from 'config';
+import { PluginsListForAppModal } from './PluginsListForAppModal';
 
 export function AppModal({
   closeModal,
@@ -212,11 +211,7 @@ export function AppModal({
           </div>
           {dependentPluginsForTemplate && dependentPluginsForTemplate.length >= 1 && (
             <div onClick={(e) => e.stopPropagation()}>
-              <div className="text-default tj-text-xsm">Install marketplace plugins used in this app</div>
-              <span className="text-placeholder tj-text-xxsm">
-                Following plugins will be installed and their respective queries in the app will be created
-              </span>
-              <PluginsList
+              <PluginsListForAppModal
                 dependentPluginsForTemplate={dependentPluginsForTemplate}
                 dependentPluginsDetail={dependentPluginsDetail}
               />
@@ -227,55 +222,3 @@ export function AppModal({
     </Modal>
   );
 }
-
-const PluginsList = ({ dependentPluginsForTemplate, dependentPluginsDetail }) => {
-  const [isExpanded, setIsExpanded] = useState(false);
-
-  const toggleExpanded = () => {
-    setIsExpanded(!isExpanded);
-  };
-
-  return (
-    <div
-      className="w-full"
-      onClick={(e) => {
-        e.stopPropagation();
-      }}
-    >
-      <span onClick={toggleExpanded}>
-        <span className="mr-1 tj-text-xsm text-default font-weight-500">View all plugins</span>
-        {isExpanded ? (
-          <SolidIcon name="arrowDownTriangle" width="12" />
-        ) : (
-          <SolidIcon name="arrowUpTriangle" width="12" />
-        )}
-      </span>
-
-      {isExpanded && dependentPluginsForTemplate && dependentPluginsForTemplate.length > 0 && (
-        <div
-          style={{
-            minHeight: '20px',
-            maxHeight: '150px',
-            overflowY: 'auto',
-            borderLeft: '1px solid var(--border-weak)',
-          }}
-        >
-          {dependentPluginsForTemplate.map((plugin, index) => {
-            const pluginsName = dependentPluginsDetail[plugin].name || plugin;
-            const iconSrc = `${config.TOOLJET_MARKETPLACE_URL}/marketplace-assets/${plugin}/lib/icon.svg`;
-            return (
-              <div
-                key={`${pluginsName}-${index}`}
-                className="d-flex custom-gap-6 flex-row align-items-center"
-                style={{ padding: '8px 7px' }}
-              >
-                <img height="15" width="15" src={iconSrc} />
-                <span className="tj-text-xsm text-default">{pluginsName}</span>
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-};
